pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "SpectraSpark"
dynamic = ["version"]
authors = [
    {name="Junnosuke Oishi"},
]
description = "X-ray scattering and absorption profile analysis tools"
# license = "MIT"
requires-python = ">=3.8"
dependencies = [
    "numpy >= 1.25",
    "scipy >= 1.0",
    "opencv-python-headless >= 4.0",
    "matplotlib >= 3.7",
    "pyqt5 >= 5.15",
    "tqdm >= 4.66",
    "pybind11 >= 2.11",
    "numba >= 0.60.0",
    "xraylarch >= 0.9.78",
    "xraydb >= 4.5.4",
]

[project.optional-dependencies]
dev = [
    "jupyter >= 1.0.0",
    "ruff >= 0.6.3",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
exclude = []

[tool.setuptools.dynamic]
version = {attr="SpectraSpark.__version__"}